+++ Classes/Feed/CategoryAwareInterface.php
@@ -1,0 +1,20 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of the "feed_generator" extension for TYPO3 CMS.
+ *
+ * For the full copyright and license information, please read the
+ * LICENSE.txt file that was distributed with this source code.
+ */
+
+namespace Brotkrueml\FeedGenerator\Feed;
+
+interface CategoryAwareInterface
+{
+    /**
+     * @return CategoryInterface[]
+     */
+    public function getCategories(): array;
+}