+++ CHANGELOG.md
@@ -6,13 +6,16 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+- Categories can be added to a feed
+
 ### Changed
 - debril/feed-io removed in favour of laminas/laminas-feed
 
 ### Removed
 - XSL stylesheets
 - JSON feeds
-- Medias in items
+- Media in items
 
 ## [0.3.0] - 2022-06-20